requirements.txt
# sYsLens v2 — Requirements
# Install: pip install -r requirements.txt

# UI
streamlit>=1.36.0

# Data validation
pydantic>=2.7.0
python-dotenv>=1.0.0

# LLM providers (openai SDK covers OpenAI + Groq + Azure)
openai>=1.35.0
anthropic>=0.30.0

# Data processing (inherited from original requirements.txt)
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
numpy>=1.26.0

# Visualization
plotly>=5.22.0

# PDF extraction (resume: "PDFs — text + table extraction")
pdfplumber>=0.11.0

# OCR image extraction (resume: "images — OCR-based extraction")
# Requires Tesseract installed on the system:
#   Ubuntu/Debian: sudo apt-get install tesseract-ocr
#   macOS:         brew install tesseract
#   Windows:       https://github.com/UB-Mannheim/tesseract/wiki
pytesseract>=0.3.10
Pillow>=10.0.0

# Sandbox (Docker SDK — optional, falls back to subprocess if unavailable)
docker>=7.1.0